rc.sysinit: Add /dev/shm to mtab if it was mounted early
Currently this early mount of /dev/shm happens in the udevd initscript
when there is no /dev/shm entry in /etc/fstab - in this case /dev/shm
is mounted with default options, therefore the added mtab entry will
be correct.  If /etc/fstab contains an entry for /dev/shm, it will be
mounted later together with other local filesystems.

Signed-off-by: Sergey Vlasov <vsu@altlinux.ru>