Use standalone systemd utilities if any
The system may have standalone versions of systemd utilities installed.
We give them priority. These utilities are less dependent on other
libraries.

Also if we don't want to fail if there are no systemd utilities at all.
This change makes the code more tolerant of the lack of these utilities.

Signed-off-by: Alexey Gladkov <legion@altlinux.org>